+++ gltf/src/net/mgsx/gltf/shaders/gdx-pbr.fs.glsl
@@ -389,7 +389,7 @@ void main() {
 
     // final frag color
 #ifdef GAMMA_CORRECTION
-    out_FragColor = vec4(pow(color,vec3(1.0/2.2)), baseColor.a);
+    out_FragColor = vec4(pow(color,vec3(1.0/GAMMA_CORRECTION)), baseColor.a);
 #else
     out_FragColor = vec4(color, baseColor.a);
 #endif
@@ -620,7 +620,7 @@ void main() {
     
     // final frag color
 #ifdef GAMMA_CORRECTION
-    out_FragColor = vec4(pow(color,vec3(1.0/2.2)), baseColor.a);
+    out_FragColor = vec4(pow(color,vec3(1.0/GAMMA_CORRECTION)), baseColor.a);
 #else
     out_FragColor = vec4(color, baseColor.a);
 #endif